chore: fix clippy warnings and improve code quality
- Replace or_insert_with(Default::default) with or_default()
- Remove unnecessary closures from ok_or_else -> ok_or
- Remove unnecessary u32 casts in sphere primitives
- Collapse nested if statements using let-else chains
- Fix duplicate match arm in editor_ui (unreachable pattern)
- Simplify boolean expression in preferences panel
- Use is_some_and() instead of map_or(false, ...) for clarity
- Remove unnecessary reference taking in convert_draw_list call

Reduced warnings from 129 to 107 (remaining mostly dead_code for public API).

+++ katla_gfx/src/primitives/sphere.rs
@@ -126,11 +126,11 @@ pub fn generate_sphere(radius: f32, segments: u32, rings: u32) -> (Vec<VertexPBR
     // We have (rings - 1) middle rings, and we connect adjacent rings
     for ring in 0..(rings - 2) {
         for segment in 0..segments {
-            let ring_base = (ring * (segments + 1)) as u32;
-            let next_ring_base = ((ring + 1) * (segments + 1)) as u32;
+            let ring_base = ring * (segments + 1);
+            let next_ring_base = (ring + 1) * (segments + 1);
 
-            let current = ring_base + segment as u32;
-            let next = next_ring_base + segment as u32;
+            let current = ring_base + segment;
+            let next = next_ring_base + segment;
 
             // Two triangles per quad (0,1,2) and (0,2,3)
             let v0 = current;
@@ -155,7 +155,7 @@ pub fn generate_sphere(radius: f32, segments: u32, rings: u32) -> (Vec<VertexPBR
 
     // Bottom cap: connect last ring to bottom pole vertices
     // Each pole vertex connects to its corresponding ring vertex
-    let last_ring_base = ((rings - 2) * (segments + 1)) as u32;
+    let last_ring_base = (rings - 2) * (segments + 1);
     for segment in 0..segments {
         let ring_vertex = last_ring_base + segment;
         let next_ring_vertex = last_ring_base + segment + 1;
